Moved sorting operation from Fragment to Recipe Adapter
