+++ android_studio_project/app/src/main/java/com/git_er_done/cmput301f22t06_team_project/controllers/RecipesRecyclerViewAdapter.java
@@ -13,12 +13,14 @@ import java.util.Comparator;
 import com.git_er_done.cmput301f22t06_team_project.R;
 import com.git_er_done.cmput301f22t06_team_project.RecipesRecyclerViewInterface;
 import com.git_er_done.cmput301f22t06_team_project.dbHelpers.RecipesDBHelper;
+import com.git_er_done.cmput301f22t06_team_project.models.Ingredient.Ingredient;
 import com.git_er_done.cmput301f22t06_team_project.models.Recipe;
 import com.git_er_done.cmput301f22t06_team_project.models.RecipeIngredient;
 import com.google.android.material.snackbar.Snackbar;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.ArrayList;
 
 public class RecipesRecyclerViewAdapter extends RecyclerView.Adapter<RecipesRecyclerViewAdapter.ViewHolder>{
 
@@ -192,6 +194,42 @@ public class RecipesRecyclerViewAdapter extends RecyclerView.Adapter<RecipesRecy
         mRecipes.add(recentlyDeletedRecipePosition, recentlyDeletedRecipe);
         notifyDataSetChanged();
         RecipesDBHelper.addRecipe(recentlyDeletedRecipe);
+    }
+    public void sortByTitle(){
+        Collections.sort(mRecipes, new Comparator<Recipe>(){
+            @Override
+            public int compare(Recipe lhs, Recipe rhs) {
+                return lhs.getTitle().compareTo(rhs.getTitle());
+            }
+        });
+        notifyDataSetChanged();
+    }
+    public void sortByCategory(){
+        Collections.sort(mRecipes, new Comparator<Recipe>(){
+            @Override
+            public int compare(Recipe lhs, Recipe rhs) {
+                return lhs.getCategory().compareTo(rhs.getCategory());
+            }
+        });
+        notifyDataSetChanged();
+    }
+    public void sortByServings(){
+        Collections.sort(mRecipes, new Comparator<Recipe>(){
+            @Override
+            public int compare(Recipe lhs, Recipe rhs) {
+                return lhs.getServings().compareTo(rhs.getServings());
+            }
+        });
+        notifyDataSetChanged();
+    }
+    public void sortByPrepTime(){
+        Collections.sort(mRecipes, new Comparator<Recipe>(){
+            @Override
+            public int compare(Recipe lhs, Recipe rhs) {
+                return lhs.getPrep_time().compareTo(rhs.getPrep_time());
+            }
+        });
+        notifyDataSetChanged();
     }
 }
